Componentファイルごとに分割する
