+++ resources/assets/jsx/components/ranking-body.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import RankingApi from '../js/ranking-api';
-import RankingTypes from '../js/ranking-types';
+import RankingApi from '../../js/ranking-api';
+import RankingTypes from '../../js/ranking-types';
 
 class RankingItem extends Component {
     constructor(props) {
@@ -80,7 +80,7 @@ class RankingItem extends Component {
     }
 }
 
-class RankingBody extends Component {
+export default class RankingBody extends Component {
     constructor(props) {
         super(props);
 
@@ -144,57 +144,3 @@ class RankingBody extends Component {
         );
     }
 }
-
-class RankingTypeNavigator extends Component {
-    constructor(props) {
-        super(props);
-
-        const { store } = props;
-
-        this.state = { store : store };
-
-        store.on("update", (updatedStore, updatedParameter) => {
-            // durationが変更されない限りnavvarへの変更も存在しない
-            if (updatedParameter !== "duration") {
-                return;
-            }
-
-            this.setState({ store : updatedStore })
-        });
-
-        this._getTab = this._getTab.bind(this);
-    }
-
-    _getTab(type) {
-        let item_class_name = "nav-item ranking-type-item";
-        if (this.state.store.type === type) {
-            item_class_name += " active";
-        }
-
-        let tab_title = RankingTypes.resolveRaw(type);
-
-        // 長さが不足していれば空白を挿入する
-        if (tab_title.length < 4) {
-            tab_title = tab_title.split("").join(" ");
-        }
-
-        return (
-            <li className={item_class_name} key={type}>
-                <a className="nav-link bg-primary" data-toggle="tab" onClick={ () => this.state.store.setType(type) }>{tab_title}</a>
-            </li>
-        );
-    }
-
-    render() {
-        return (
-            <ul className="nav nav-tabs" id="ranking-type-nav">
-                { RankingTypes.getAvailableTypes(this.state.store.duration).map(this._getTab) }
-            </ul>
-        );
-    }
-}
-
-export {
-    RankingBody,
-    RankingTypeNavigator
-};
