+++ resources/assets/jsx/ranking.jsx
@@ -2,10 +2,11 @@
 
 import React from "react";
 import ReactDOM from 'react-dom';
-import { RankingBody, RankingTypeNavigator } from "./ranking-components.jsx";
+
+import RankingBody from "./components/ranking-body.jsx";
+import RankingTypeNavigator from "./components/ranking-type-navigator.jsx"
+
 import rankingStore from "../js/ranking-store";
 
-(() => {
-    ReactDOM.render(<RankingBody store={rankingStore}/>, document.getElementById('ranking-container'));
-    ReactDOM.render(<RankingTypeNavigator store={rankingStore}/>, document.getElementById('ranking-type-nav'));
-})();
+ReactDOM.render(<RankingBody store={rankingStore}/>, document.getElementById('ranking-container'));
+ReactDOM.render(<RankingTypeNavigator store={rankingStore}/>, document.getElementById('ranking-type-nav'));
